rename async -> nonblocking for python 3.7 compatibility
